requirements.txt
flask
opensimplex
click
pillow